% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bin_data}
\alias{bin_data}
\title{A simulated binomial dataset.}
\format{
A list with two elements: design matrix \code{X} and \code{y}.
\describe{
  \item{X}{design matrix}
  \item{y}{response}
  ...
}
}
\usage{
data(bin_data)
}
\description{
A dataset simulated for illustrating logistic regression models.
}
\examples{
data("bin_data")
cv.fit <- cv.glmtlp(bin_data$X, bin_data$y, family = "binomial", penalty = "l1")
plot(cv.fit)

}
\keyword{datasets}
